Move CreateDatabase and DropDatabase into an utility class.
